Fixed some more cases where I missed the long -> int client id conversion.
git-svn-id: https://jmonkeyengine.googlecode.com/svn/trunk@7028 75d07b2b-3a1a-0410-a2c5-0572b91ccdca

+++ engine/src/networking/com/jme3/network/base/DefaultServer.java
@@ -64,7 +64,7 @@ public class DefaultServer implements Server
     private Kernel fast;
     private KernelAdapter fastAdapter;
     private Redispatch dispatcher = new Redispatch();
-    private Map<Long,HostedConnection> connections = new ConcurrentHashMap<Long,HostedConnection>();
+    private Map<Integer,HostedConnection> connections = new ConcurrentHashMap<Integer,HostedConnection>();
     private Map<Endpoint,HostedConnection> endpointConnections 
                             = new ConcurrentHashMap<Endpoint,HostedConnection>();
     
@@ -346,7 +346,7 @@ public class DefaultServer implements Server
             id = nextId.getAndIncrement();
         }
         
-        public long getId()
+        public int getId()
         {
             return id;
         }
